refactor(notification): 노티 활성화 기능 추가
