+++ docker-compose.backend.yml
@@ -32,7 +32,7 @@ services:
       - 3000:3000
     environment:
       - NODE_ENV=development
-      - NEXT_PUBLIC_API_URL=http://192.168.45.8:8080
+      - NEXT_PUBLIC_API_URL=http://localhost:8080
       - NEXT_PUBLIC_KAKAOMAP_APPKEY=53a909c464c1aac1af1ee43c1f03ce86
       - NEXT_PUBLIC_KAKAOMAP_APIKEY=eec1369924fcf9860f419cd6fcb0b82c
     volumes:
